requirements.txt
matplotlib>=3.8,<4
PyYAML>=6.0,<7
